feat(interface): enable float value timeout

+++ skmp/solver/myrrt_solver.py
@@ -21,7 +21,7 @@ class MyRRTConfig:
     n_max_satisfaction_trial: int = 100
     satisfaction_conf: Optional[SatisfactionConfig] = SatisfactionConfig()
     n_subgoal: int = 4  # used only when init_traj is specified
-    timeout: Optional[int] = None
+    timeout: Optional[float] = None
 
     @property
     def sample_goal_first(self) -> bool:
